+++ cardinal/server/persona.go
@@ -1,4 +1,4 @@
-package server1
+package server
 
 import (
 	"errors"
